+++ gatsby-config.js
@@ -30,6 +30,26 @@ module.exports = {
           },
         },
         {
+          resolve: `gatsby-plugin-alias-imports`,
+          options: {
+            alias: {
+              "@components": "src/components",
+              "@icons": "src/assets/icons"
+            },
+            extensions: [
+              "js",
+            ],
+          }
+        },
+        {
+          resolve: "gatsby-plugin-react-svg",
+          options: {
+            rule: {
+              include: /assets/ // See below to configure properly
+            }
+          }
+        },
+        {
             resolve: `gatsby-plugin-manifest`,
             options: {
                 name: `gatsby-starter-default`,
@@ -40,7 +60,7 @@ module.exports = {
                 // https://css-tricks.com/meta-theme-color-and-trickery/
                 // theme_color: `#663399`,
                 display: `minimal-ui`,
-                icon: `src/images/icon.png`, // This path is relative to the root of the site.
+                icon: `src/assets/icons/favicon.png`, // This path is relative to the root of the site.
             },
         },
         // this (optional) plugin enables Progressive Web App + Offline functionality
