+++ package.json
@@ -6,17 +6,18 @@
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "dependencies": {
     "autoprefixer": "^10.4.7",
-    "bootstrap": "^5.1.3",
     "browserslist": "^4.20.3",
     "firebase": "^9.8.1",
     "foundation-sites": "^6.7.4",
     "gatsby": "^4.15.0",
+    "gatsby-plugin-alias-imports": "^1.0.5",
     "gatsby-plugin-gatsby-cloud": "^4.15.0",
     "gatsby-plugin-image": "^2.15.0",
     "gatsby-plugin-manifest": "^4.15.0",
     "gatsby-plugin-offline": "^5.15.0",
     "gatsby-plugin-postcss": "^5.15.0",
     "gatsby-plugin-react-helmet": "^5.15.0",
+    "gatsby-plugin-react-svg": "^3.1.0",
     "gatsby-plugin-sass": "^5.15.0",
     "gatsby-plugin-sharp": "^4.15.0",
     "gatsby-source-filesystem": "^4.15.0",
@@ -26,7 +27,6 @@
     "postcss-cli": "^9.1.0",
     "prop-types": "^15.8.1",
     "react": "^17.0.1",
-    "react-bootstrap": "^2.4.0",
     "react-dom": "^17.0.1",
     "react-firebase-hooks": "^5.0.3",
     "react-helmet": "^6.1.0",
