+++ src/components/header/header.scss
@@ -1,0 +1,5 @@
+#masthead {
+  .inner {
+    background: red;
+  }
+}
