+++ src/pages/page-2.js
@@ -1,16 +1,14 @@
 import * as React from "react"
 import { Link } from "gatsby"
-
-import Layout from "../components/layout"
 import Seo from "../components/seo"
 
 const SecondPage = () => (
-  <Layout>
+  <div>
     <Seo title="Page two" />
     <h1>Hi from the second page</h1>
     <p>Welcome to page 2</p>
     <Link to="/">Go back to the homepage</Link>
-  </Layout>
+  </div>
 )
 
 export default SecondPage
