+++ src/templates/using-dsg.js
@@ -1,11 +1,10 @@
 import * as React from "react"
 import { Link } from "gatsby"
 
-import Layout from "../components/layout"
 import Seo from "../components/seo"
 
 const UsingDSG = ({data}) => (
-    <Layout>
+    <div>
         <Seo title="Using DSG" />
         <h1>
             Hello from a <b>DSG Page</b>
@@ -20,7 +19,7 @@ const UsingDSG = ({data}) => (
             .
         </p>
         <Link to="/">Go back to the homepage</Link>
-    </Layout>
+    </div>
 )
 
 export default UsingDSG
